至少保留最近的3个版本

+++ servers/gs-app/scripts/deploy-gs-vue.sh
@@ -56,14 +56,18 @@ else
     # PM2重启项目
     pm2 restart gs-vue-ssr
 
-    # 删除过期的目录(默认删除3天前的过期目录)
+    # 删除过期的目录(默认删除3天前的过期目录), 并保存最新 3个文件夹
+    reservedNum=3
     currentDate=`date +%s`
-    let isDeleted=24*3600*3
+    let isDeleted=24*3600*7
     for file in $project_path/version/*; do
-        createdTime=`stat -c %Y $file`
-        if [ $[ $currentDate - $createdTime ] -gt $isDeleted ]; then
-            echo "Delete File: $file"
-            /bin/rm -rf $file
+        folderNum=$(ls -l $project_path/version | grep 'source' | wc -l)
+        if [ $folderNum -gt $reservedNum ]; then
+            createdTime=`stat -c %Y $file`
+            if [ $[ $currentDate - $createdTime ] -gt $isDeleted ]; then
+                echo "Delete File: $file"
+                /bin/rm -rf $file
+            fi
         fi
     done
 fi
